CI: Remove deprecated files
